+++ test/Spec.hs
@@ -14,5 +14,4 @@ import Test.Tasty (defaultMain)
 import Test.Relude.Property (hedgehogTestTree)
 
 main :: IO ()
-main = do
-    defaultMain hedgehogTestTree
+main = defaultMain hedgehogTestTree