+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+0.3.1
+=====
+* [#81](https://github.com/kowainik/relude/issues/81):
+  Add `asumMap` to `Foldable` functions.
+
 0.3.0
 =====
 
@@ -35,8 +40,7 @@
   Add `under2` and `underF2` functions to `Relude.Extra.Newtype`.
 * [#60](https://github.com/kowainik/relude/issues/60):
   Add `hoistMaybe` and `hoistEither` functions.
-* [#81](https://github.com/kowainik/relude/issues/81):
-  Add `asumMap` to `Foldable` functions.
+
 
 0.2.0
 =====